+++ packages/ipa-tools/CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2026-08-20
+
 ### Changed
 
-- Document `createInference` fallbacks with `ipa-prompt-api-fallback` object example ([#20](https://github.com/SamSamskies/inference-provider-api/issues/20)).
+- Document `createInference` fallbacks with `ipa-prompt-api-fallback` object example ([#20](https://github.com/SamSamskies/inference-provider-api/issues/20) phase 3).
 
 ## [0.4.0] - 2026-08-19